Tangle

Tangle.KnotParameter — Type

KnotParameter is a floating point number modulo 1.

Tangle.Line — Type

A Line is represented by two points on it. For a line segment, these points are the end points.

Tangle.Line — Method

A Line can be treated as a parametric function to identify some point on the line.

Tangle.Loop — Type

Loop

A Loop represents a mathematical knot.

In one sense, it is a function in some parameter p (a KnotParameter) that returns a point in R3 for p.

loop(0.0) == loop(1.0)

where loop is a Loop.

As a means of defining and representing knots, a Loop has a sorted set of "points of interest". Each point of interest is just a specific value for the parameter t.

Tangle.Loop — Method

(::Loop)(at)

Return a three element vector of the X, Y and Z coordinates of the point of the Loop corresponding to `at'.

This function should be continuous and defined by the specified points of interest.

Tangle.LoopSegmentsIterator — Type

LoopSegmentsIterator(::Loop)

Return an iterator for iterating over the segments of a Loop.

Each segment is a two element tuple of successive points of interest of Loop.

Tangle.PointOfInterest — Type

PointOfInterest

PointOfInterest represents a point on a Loop for which a location has been specified.

No two points of interest should have the same KnotParameter widthin a Loop.

No two points of interest should have equal coordinates within the same mathematical link.

Currently, no code enforces these rules.

Tangle.divide_interval — Method

divide_interval(from::KnotParameter, to::KnotParameter, count::Integer

Return count KnotParameters evenly spaced between from and to.

Tangle.find_poi — Method

find_poi(predicate, ::Loop)

Return the PointOfInterest from the Loop which satisfies predicate.

Tangle.grab — Method

grab(::Loop, ::KnotParameter, label, delta)

Return a new Loop with a point of interest added at the specified KnotParameter with the specified symbol.

The new point of interest will be displaced by delta "outward".

label provides a way to label or explain the reason for grabbing.

Tangle.grab — Method

grab(loop::Loop, at::KnotParameter, delta, label::Symbol)

Return a new loop with a point of interest added between the one labeled after and the next point of interest after it.

The new point of interest will be displaced by delta "outward".

label provides a way to label or explain the reason for grabbing.

Tangle.next — Method

next(loop::Loop, kp::KnotParameter)

If kp is the KnotParameter of some PointOfInterest in loop then return the next PointOfInterest in that loop.

Tangle.parallel — Method

parallel(line1::Line, line2::Line)::Bool

Return true if the two lines are parallel.

Tangle.previous — Method

previous(loop::Loop, kp::KnotParameter)

If kp is the KnotParameter of some PointOfInterest in loop then return the next PointOfInterest in that loop. Otherwise, return nothing.

Tangle.proximal_points — Method

proximal_points(line1::Line, line2::Line)

Return the two points (as 3 element Vectors) which are the point on line1 closest to line2 and the point on line2 closest to line1 respectively.

Tangle.reidemeisterTwist — Method

reidemeisterTwist(loop::Loop, handle::PointOfInterest, twist::Twist)

Return a new loop by applying the Reidermeister I (twist) operation to loop at handle.

twist determines the direstion of twist.

handle should be the result of a grab operation.

Tangle.symbolic_vector — Method

symbolic_vector(name::Symbol, n::Integer)

Define n Symbolics variables whose names are based on name and return a vector of those variables.

Tangle.vsubs — Method

vsubs(v::Vector{Symbolics.Num}, values...)

Return a list of Pairs suitable for constructing a Dict which is suitable for passing to Symbolics.substitute, which substitues each of values with the corresponding variable in v.

Tangle.was — Method

was(::Grab)

Return the PointOfInterest that this operation grabbed from.