Tangle

Tangle is an attempt at software for describing, modeling and working with knots.

Representation

The basic data structure is a Loop. Among other things, a Loop is a function that maps a KnotParameter to a point in three dimensional space.

A KnotParameter is a floating poingt number ranging from 0.0 (MIN_KnotParameter) to 1.0 (MAX_KnotParameter). Arithmetic on KnotParameters is performed modulo 1.0.

A loop is defined by a sequence of ":points of interest". Each PointOfIntrerest assiciates a KnotParameter with a point in space.

Tangle.KnotParameter — Type

KnotParameter(param::Float16)

KnotParameter is a floating point value modulo 1. It represents the parameter for a point on a parametric curve.

KnotParameters can be added and subtracted. They can be multiplied by or divided by numbers.

Tangle.divide_interval — Function

divide_interval(from::KnotParameter, to::KnotParameter, count::Integer

Return count KnotParameters evenly spaced between from and to.

Tangle.PointOfInterest — Type

PointOfInterest

PointOfInterest represents a point on a Loop for which a location has been specified.

No two points of interest should have the same KnotParameter widthin a Loop.

No two points of interest should have equal coordinates within the same mathematical link.

Currently, no code enforces these rules.

Tangle.PointsOfInterest — Type

PointsOfInterest is a sequence whose elements are PointsOfInterest objects.

Tangle.Loop — Type

Loop

A Loop represents a mathematical knot.

In one sense, it is a function in some parameter p (a KnotParameter) that returns a point in R3 for p.

loop(0.0) == loop(1.0)

where loop is a Loop.

As a means of defining and representing knots, a Loop has a sorted set of "points of interest". Each point of interest is just a specific value for the parameter t.

Tangle.find_poi — Function

find_poi(predicate, ::Loop)

Return the PointOfInterest from the Loop which satisfies predicate.

Tangle.next — Function

next(loop::Loop, kp::KnotParameter)

If kp is the KnotParameter of some PointOfInterest in loop then return the next PointOfInterest in that loop.

Tangle.previous — Function

previous(loop::Loop, kp::KnotParameter)

If kp is the KnotParameter of some PointOfInterest in loop then return the next PointOfInterest in that loop. Otherwise, return nothing.

Tangle.before — Function

before(loop::Loop, label)
before(loop::Loop, poi::PointOfInterest)::KnotParameter

Return a KnotParameter that comes before the specified PointOfInterest but after any previous points of interest.

Tangle.after — Function

after(loop::Loop, label)
after(loop::Loop, poi::PointOfInterest)::KnotParameter

Return a KnotParameter that comes after the specified PointOfInterest but before any succeeding points of interest.

Operations

Tangle.grab — Function

grab(::Loop, ::KnotParameter, label, delta)

Return a new Loop with a point of interest added at the specified KnotParameter with the specified symbol.

The new point of interest will be displaced by delta "outward".

label provides a way to label or explain the reason for grabbing.

Tangle.was — Function

was(::Grab)

Return the PointOfInterest that this operation grabbed from.

Tangle.reidemeisterTwist — Function

reidemeisterTwist(loop::Loop, handle::PointOfInterest, twist::Twist)

Return a new loop by applying the Reidermeister I (twist) operation to loop at handle.

twist determines the direstion of twist.

handle should be the result of a grab operation.

The diameter of the loop is based on the delta of handle's Grab operation.

Geometry

Tangle.center — Function

center(points)

Return the point at the center of the collection of points.

Tangle.Line — Type

A Line is represented by two points on it. For a line segment, these points are the end points.

Tangle.direction_vector — Function

direction_vector(::Line)

Return a direction vector for the Line.

Tangle.unit_direction_vector — Function

unit_direction_vector(::Line)

Return the unit direction vector for the Line.

Tangle.parallel — Function

parallel(line1::Line, line2::Line)::Bool

Return true if the two lines are parallel.

Tangle.point_on_line — Function

point_on_line(point, line::Line)::Bool

Return true if line intersects point.

Tangle.point_in_segment — Function

point_in_segment(point, segment::Line)

If point is on the specified Line and lies within the defining points of that Line then return the parameter p such that line(p) == point, otherwise return nothing.

Tangle.proximal_points — Function

proximal_points(line1::Line, line2::Line)

Return the two points (as 3 element Vectors) which are the point on line1 closest to line2 and the point on line2 closest to line1 respectively.

Symbolics.jl

Portions of Tangle use Symbolics.jl. Some utilities have been defined to make it easier to use vectors in Symbolics.jl.

Tangle.symbolic_vector — Function

symbolic_vector(name::Symbol, n::Integer)

Define n Symbolics variables whose names are based on name and return a vector of those variables.

Tangle.vsubs — Function

vsubs(v::Vector{Symbolics.Num}, values...)

Return a list of Pairs suitable for constructing a Dict which is suitable for passing to Symbolics.substitute, which substitues each of values with the corresponding variable in v.